refactor(arena): agentkb as single source of truth for examples + schemas
Two duplications collapse onto agentkb:
- Move the example renderer into agentkb (RenderExample / ExampleLoader);
  the examples-show command and the MCP show_example tool now share one
  definition instead of two copies.
- Add a go:generate directive that mirrors schemas/v1alpha1 into the agentkb
  embed, wired into `make schemas`. The byte-parity test guards drift; this
  gives a one-command regeneration path. The same mechanism will later
  generate the authoring skill from concepts.
