+++ tools/arena/cmd/promptarena/examples_cmd.go
@@ -13,12 +13,6 @@ import (
 	"github.com/AltairaLabs/PromptKit/tools/arena/templates"
 )
 
-// exampleLoader is the subset of templates.Loader that examples show needs.
-type exampleLoader interface {
-	Load(name string) (*templates.Template, error)
-	ReadTemplateFile(templateName, filePath string) ([]byte, error)
-}
-
 const examplesListUse = "list"
 
 var (
@@ -102,65 +96,13 @@ func writeExamplesList(w io.Writer, tag string, asJSON bool) error {
 	return tw.Flush()
 }
 
-func writeExampleShow(w io.Writer, loader exampleLoader, name string) error {
-	cat, err := agentkb.LoadCatalog()
+func writeExampleShow(w io.Writer, loader agentkb.ExampleLoader, name string) error {
+	text, err := agentkb.RenderExample(loader, name)
 	if err != nil {
 		return err
 	}
-
-	var entry *agentkb.CatalogEntry
-	for i := range cat.Entries {
-		if cat.Entries[i].Name == name {
-			entry = &cat.Entries[i]
-			break
-		}
-	}
-	if entry == nil {
-		return fmt.Errorf("unknown example %q (try 'promptarena examples list')", name)
-	}
-
-	loadName := entry.Source
-	if idx := strings.IndexByte(loadName, ':'); idx >= 0 {
-		loadName = loadName[idx+1:]
-	}
-
-	tmpl, err := loader.Load(loadName)
-	if err != nil {
-		return fmt.Errorf("load example %q: %w", name, err)
-	}
-
-	if _, err := fmt.Fprintf(w, "# %s\n\n%s\n\nFiles:\n", entry.Name, entry.Description); err != nil {
-		return err
-	}
-	for i := range tmpl.Spec.Files {
-		f := &tmpl.Spec.Files[i]
-		if _, err := fmt.Fprintf(w, "\n--- %s ---\n", f.Path); err != nil {
-			return err
-		}
-		body := fileBody(loader, tmpl.Metadata.Name, f)
-		if _, err := fmt.Fprintln(w, body); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// fileBody returns the displayable body for a template file: inline content, the
-// embedded template source, or a placeholder when the body lives elsewhere
-// (external/remote source).
-func fileBody(loader exampleLoader, templateName string, f *templates.FileSpec) string {
-	switch {
-	case f.Content != "":
-		return f.Content
-	case f.Template != "":
-		data, err := loader.ReadTemplateFile(templateName, f.Template)
-		if err != nil {
-			return fmt.Sprintf("(template body unavailable: %v)", err)
-		}
-		return string(data)
-	default:
-		return "(external source — fetch the example to see this file)"
-	}
+	_, err = io.WriteString(w, text)
+	return err
 }
 
 func containsString(haystack []string, needle string) bool {
